ugly merge

+++ plugins/rocksdb/src/main/java/org/hyperledger/besu/plugin/services/storage/rocksdb/segmented/RocksDBSnapshotTransaction.java
@@ -22,6 +22,7 @@ import org.hyperledger.besu.plugin.services.storage.rocksdb.RocksDBMetrics;
 import org.hyperledger.besu.plugin.services.storage.rocksdb.RocksDbIterator;
 
 import java.util.Optional;
+import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.stream.Stream;
 
 import org.apache.commons.lang3.tuple.Pair;
@@ -35,6 +36,7 @@ import org.rocksdb.WriteOptions;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+/** The Rocks db snapshot transaction. */
 public class RocksDBSnapshotTransaction implements KeyValueStorageTransaction, AutoCloseable {
   private static final Logger LOG = LoggerFactory.getLogger(RocksDBSnapshotTransaction.class);
   private static final String NO_SPACE_LEFT_ON_DEVICE = "No space left on device";
@@ -46,7 +48,15 @@ public class RocksDBSnapshotTransaction implements KeyValueStorageTransaction, A
   private final RocksDBSnapshot snapshot;
   private final WriteOptions writeOptions;
   private final ReadOptions readOptions;
+  private final AtomicBoolean isClosed = new AtomicBoolean(false);
 
+  /**
+   * Instantiates a new RocksDb snapshot transaction.
+   *
+   * @param db the db
+   * @param columnFamilyHandle the column family handle
+   * @param metrics the metrics
+   */
   RocksDBSnapshotTransaction(
       final OptimisticTransactionDB db,
       final ColumnFamilyHandle columnFamilyHandle,
@@ -76,7 +86,18 @@ public class RocksDBSnapshotTransaction implements KeyValueStorageTransaction, A
     this.snapTx = snapTx;
   }
 
+  /**
+   * Get data against given key.
+   *
+   * @param key the key
+   * @return the optional data
+   */
   public Optional<byte[]> get(final byte[] key) {
+    if (isClosed.get()) {
+      LOG.debug("Attempted to access closed snapshot");
+      return Optional.empty();
+    }
+
     try (final OperationTimer.TimingContext ignored = metrics.getReadLatency().startTimer()) {
       return Optional.ofNullable(snapTx.get(columnFamilyHandle, readOptions, key));
     } catch (final RocksDBException e) {
@@ -86,6 +107,11 @@ public class RocksDBSnapshotTransaction implements KeyValueStorageTransaction, A
 
   @Override
   public void put(final byte[] key, final byte[] value) {
+    if (isClosed.get()) {
+      LOG.debug("Attempted to access closed snapshot");
+      return;
+    }
+
     try (final OperationTimer.TimingContext ignored = metrics.getWriteLatency().startTimer()) {
       snapTx.put(columnFamilyHandle, key, value);
     } catch (final RocksDBException e) {
@@ -99,6 +125,10 @@ public class RocksDBSnapshotTransaction implements KeyValueStorageTransaction, A
 
   @Override
   public void remove(final byte[] key) {
+    if (isClosed.get()) {
+      LOG.debug("Attempted to access closed snapshot");
+      return;
+    }
     try (final OperationTimer.TimingContext ignored = metrics.getRemoveLatency().startTimer()) {
       snapTx.delete(columnFamilyHandle, key);
     } catch (final RocksDBException e) {
@@ -110,12 +140,22 @@ public class RocksDBSnapshotTransaction implements KeyValueStorageTransaction, A
     }
   }
 
+  /**
+   * Stream.
+   *
+   * @return the stream
+   */
   public Stream<Pair<byte[], byte[]>> stream() {
     final RocksIterator rocksIterator = db.newIterator(columnFamilyHandle, readOptions);
     rocksIterator.seekToFirst();
     return RocksDbIterator.create(rocksIterator).toStream();
   }
 
+  /**
+   * Stream keys.
+   *
+   * @return the stream
+   */
   public Stream<byte[]> streamKeys() {
     final RocksIterator rocksIterator = db.newIterator(columnFamilyHandle, readOptions);
     rocksIterator.seekToFirst();
@@ -124,8 +164,7 @@ public class RocksDBSnapshotTransaction implements KeyValueStorageTransaction, A
 
   @Override
   public void commit() throws StorageException {
-    // no-op or throw?
-    throw new UnsupportedOperationException("RocksDBSnapshotTransaction does not support commit");
+    // no-op
   }
 
   @Override
@@ -144,7 +183,15 @@ public class RocksDBSnapshotTransaction implements KeyValueStorageTransaction, A
     }
   }
 
+  /**
+   * Copy.
+   *
+   * @return the rocks db snapshot transaction
+   */
   public RocksDBSnapshotTransaction copy() {
+    if (isClosed.get()) {
+      throw new StorageException("Snapshot already closed");
+    }
     try {
       var copyReadOptions = new ReadOptions().setSnapshot(snapshot.markAndUseSnapshot());
       var copySnapTx = db.beginTransaction(writeOptions);
@@ -164,5 +211,6 @@ public class RocksDBSnapshotTransaction implements KeyValueStorageTransaction, A
     writeOptions.close();
     readOptions.close();
     snapshot.unMarkSnapshot();
+    isClosed.set(true);
   }
 }
